Create container factory
Reference #77
